+++ src/services/paper.service.ts
@@ -1,13 +1,17 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { PrismaService } from './prisma.service';
 import { Paper } from '@prisma/client';
 import { createReadStream } from 'fs';
 import { CreatePaperDto } from 'src/dto/paper/create-paper.dto';
 import { UpdatePaperDto } from 'src/dto/paper/update-paper.dto';
+import { GetPaperDto } from 'src/dto/paper/get-paper.dto';
+import { FileService } from './file.service';
 
 @Injectable()
 export class PaperService {
-  constructor(private prisma: PrismaService) {}
+  constructor(private prisma: PrismaService,
+    private fileService: FileService
+  ) {}
 
   async createPaper(path?: string, data?: CreatePaperDto) : Promise<Paper> {
     let fileData: Buffer;
@@ -25,11 +29,13 @@ export class PaperService {
       data: {
         name: data?.name,
         description: data?.description,
-        applicationSphere: data?.applicationSphere,
+        applicationSphere: data?.applicationSphere?.join('@#$'),
         categoryId: data?.categoryId,
         picture: fileData,
       },
     });
+
+    await this.fileService.deleteFile(path);
 
     return paper;
   }
@@ -53,30 +59,63 @@ export class PaperService {
       data: {
         name: data?.name,
         description: data?.description,
-        applicationSphere: data?.applicationSphere,
+        applicationSphere: data?.applicationSphere?.join('@#$'),
         categoryId: data?.categoryId,
         picture: fileData,
       },
     });
 
+    await this.fileService.deleteFile(path);
+    
     return paper;
   }
 
   async deletePaper(id: string): Promise<Paper> {
+    const cargos = await this.prisma.cargo.findMany({
+      where: {paperId: id},
+      select: {
+        id: true
+      }
+    });
+
+    if(cargos.length > 0) {
+      throw new HttpException('Раздел содержит товары!', HttpStatus.BAD_REQUEST);
+    }
+
     return await this.prisma.paper.delete({
       where: {id}
     });
   }
 
-  async getById(id: string): Promise<Paper | null> {
-    return await this.prisma.paper.findUnique({
+  async getById(id: string): Promise<GetPaperDto | null> {
+    const paper = await this.prisma.paper.findUnique({
       where: {id},
     });
+
+    return {
+      id: paper?.id,
+      name: paper?.name,
+      description: paper?.description,
+      applicationSphere: paper?.applicationSphere?.split('@#$'),
+      categoryId: paper?.categoryId,
+      picture: paper?.picture
+    }
   }
 
-  async getAll(categoryId?: string): Promise<Paper[]> {
-    return await this.prisma.paper.findMany({
+  async getAll(categoryId?: string): Promise<GetPaperDto[]> {
+    const papers = await this.prisma.paper.findMany({
       where:{categoryId}
+    });
+
+    return papers.map(paper =>  {
+      return {
+        id: paper?.id,
+        name: paper?.name,
+        description: paper?.description,
+        applicationSphere: paper?.applicationSphere?.split('@#$'),
+        categoryId: paper?.categoryId,
+        picture: paper?.picture
+      }
     });
   }
 }
